refactor: extract orderable config helpers into utils
Split the orderable setup and reorder endpoint internals into focused utility functions with TSDoc so index orchestration stays minimal and easier to review, while preserving join-scope ordering behavior.
