+++ packages/payload/src/config/orderable/utils/createJSONResponse.ts
@@ -1,0 +1,9 @@
+/**
+ * Creates a JSON `Response` with consistent headers and status code.
+ */
+export function createJSONResponse(body: unknown, status: number): Response {
+  return new Response(JSON.stringify(body), {
+    headers: { 'Content-Type': 'application/json' },
+    status,
+  })
+}
